PBUILD changes for OMI.
Proper PBUILD is now in DEV branch, not mainline. UPDATE YOUR TF
ENLISTMENT!

A number of changes were made that may require changes to your .pbuild
files:

1) Removed old "cache" setting (this was historical from before we
used 'configure' scripts).  This has been replaced with --debug /
--nodebug command line qualifiers, along with a default "Debug"
setting.

2) In your .pbuild file, host lines with no project field used to
assume OM.  This is now no longer valid; project MUST be specified on
host lines or an error will result.

3) Branches were allowed to be empty before (in which case no branch
specification could appear on host lines).  This is no longer valid.

In short, I worked hard to maintain backwards compatibility with old
.pbuild files in the past.  I'm now biting the bullet (and so are
you).  Update your .pbuild files to the latest format.

Removed --full qualifier.  That is now assumed.

Because of removal of cache setting, changed the way that
Debug/NoDebug is determined: Added new setting (Debug or NoDebug) to
control debug builds.  Default setting: Debug.  Added new command
line qualifiers (--debug, --nodebug) to explicitly override the
settings flag.  Order of priority: command line, .pbuild settings
flag, PBUILD code

Added new setting: CheckValidity.  Default is on.  If set to
NoCheckValidity, then your .pbuild file is not checked for validity
each time you edit your .pbuild file.  This can cause build failures,
and is intended for experienced users only.

Updated config_sample file so that it will work appropriately with
PBUILD due to above changes.

Updated README file in PBUILD directory for these changes as
appropriate.

Internal changes to PBUILD to track host based on host names rather
than tags.  This allows us to support building multiple branches
simultaneously (down the road).

Added support for project "nip" (stands for non-integrated project).
This is what the OMI project will use.  To support this, added new
qualifier: --container.  The --container=<file> should specify a
filename that is an uncompressed .tar file.  With that, we make a
bundle, copy that to the destination system, unpack the .tar file, and
do the build.  Note that this is currently hard-coded for the OMI
project (i.e. ./configure --dev), but I plan to make this configurable
via the .pbuild file in a future checkin.

In total, OMI support is comprised of the following: With --container
(required for a regular build), we make bundles and use them
remotely.  Without --container, you must specify --command.  This is
for fast execution of remote command.  With --container=<file>
--target='none;<command>', we will unpack bundle then execute
specified command.

These changes allow for execution of a quick command (to fix a
problem on the systems, say), as well as execution of commands
within the context of building your project.