Abort a build if we determine clock skew has occurred.
I'm not 100% certain if this is "good" or not. Most of the time I've
seen clock skew, it causes a build to loop forever until a DEV notices
and then kills it. With this change, PBUILD will detect the warning
from make and abort automatically. Generally, death of the ssh process
(that launches the build) appears to kill the remove build, but I do
have a warning at line 377 for the Developer to validate this.

The reason I'm not 100% certain if this is "good" or not" If there
were clock skews that didn't result in an infinite build loop, this
will cause the build to abort. I have no idea how often this came up.

I propose checking this in, then collecting data over time. Once we've
collected data, there are a variety of options:

1) Revert the change,
2) Keep the change as is,
3) Keep the change as is but add an option to disable clock skew
   detection, and
4) Keep the change as is but add code to automatically try and "fix"
   the build and restart it (find <dirpaths> -exec touch {} \;)